requirements.txt
torch
torchvision
pytorch-ignite
pyyaml